refactor : Change project api version and remove  unused code
